Begin refactoring for different architectures
