+++ src/lib.rs
@@ -1,6 +1,7 @@
 // #![allow(dead_code)]
 #[macro_use]
 extern crate lazy_static;
+pub mod arch;
 pub mod assembler;
 pub mod instruction;
 pub mod isa;
